Exception message fix
I set this up at the beginning of the project and totally expected it to
work fully. This was when I first started learning C++ and was a "Create
and forget" setup. It was definitely having issues along the development
process, but I didn't think it was an issue with this. Now that I'm more
familiar, I said "well no shit, sherlock. This is obviously wrong".