Removed empty and duplicate materials
Duplicate materials were somehow intentional, I'm not sure what the
logic was at the time. Now duplicates are combined into 1 material.

Level objects can override materials on the mesh with a name array. The
override material array needs to be the same size as submeshes within
the mesh file. To fill a space in the array but not override, the empty
quotes "" is used. The extractor no longer replaces the materal name
with empty quotes.