Rotation mostly complete, now initial transforms
I believe rotation is working fully, and the Tenno liset model was just
a bad example. It seems to work fine with the current algorithm. I've
fixed the code to allow direct application to models to further test
correctness, and am realizing the initial transformations are not being
applied properly.