Binary-Reader and LotusLib huge refactor
Binary-Reader is now header-only, LotusLib operations now use
higher-level classes for much cleaner data access.

1. Binary-Reader has been moved to its own namespace
2. LotusLib has a method getFile() that retrieves the Commonheader, and
   data from the 3 pacakges (H, B, F). Unsurprisingly this took a very
   long time to fully implement and is a majority of this commit.

The nasty code was bothering me for so long, and if I plan on making
extractors for other games, an wanted to test my theory that helper
methods are a good addition. I was right, this code is so much cleaner.