Removed Ensmallening struct
This was intended to be a "future-proof" feature so I could account for
each stage of the Ensmallening. Now after having written many formats, I
realize this is a hinderance more than it's useful. I only used it once
in the texture reader.

Anyway, the game WARFRAME_PE can be used in the future. In-between each
fo the ensmallening stages? Nah, not dealing with that.