Models export with pos/rot/scale applied at import
I realized levelstatic models could have their size drastically reduced
with deduplication, but this required pos/rot/scale to be applied on
glTF nodes, not pre-applied to vertices. To make this a completed (good)
implementation, changes will need to be made to how Models and Levels
(non-static) are exported. But that will come later. Probably.
Hopefully.