Added a search field
Added a search field, button to clear it and now after the tree loads in it is indexed for faster search. Also added some 300ms padding before search activates for culling partial searches. After the search is cleared it'll return the tree to normal + leaves the expanded items from the latest search.